# docs/openapi-v3.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini",
    "title": "Mecanismo de Busca API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Prefeitura do Rio de Janeiro",
      "url": "https://prefeitura.rio",
      "email": "contato@prefeitura.rio"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/v1/admin/services": {
      "get": {
        "description": "Lista serviços com paginação e filtros opcionais. Cada serviço na resposta inclui campos plaintext gerados automaticamente (resumo_plaintext, resultado_solicitacao_plaintext, descricao_completa_plaintext, documentos_necessarios_plaintext, instrucoes_solicitante_plaintext) que removem toda formatação markdown.",
        "tags": [
          "admin"
        ],
        "summary": "Lista serviços com paginação e filtros",
        "parameters": [
          {
            "description": "Página",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "description": "Resultados por página",
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "description": "Status do serviço (0=Draft, 1=Published)",
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Filtrar por autor",
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filtrar por tema geral",
            "name": "tema_geral",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filtrar por aguardando aprovação",
            "name": "awaiting_approval",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Filtrar por serviços gratuitos",
            "name": "is_free",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Filtrar por data de publicação (timestamp)",
            "name": "published_at",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Filtrar por nome do serviço",
            "name": "nome_servico",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Campo para filtro dinâmico",
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Valor para filtro dinâmico (usado com field)",
            "name": "value",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.PrefRioServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Cria um novo serviço na collection prefrio_services_base. A resposta inclui campos plaintext gerados automaticamente (resumo_plaintext, resultado_solicitacao_plaintext, descricao_completa_plaintext, documentos_necessarios_plaintext, instrucoes_solicitante_plaintext) que removem toda formatação markdown.",
        "tags": [
          "admin"
        ],
        "summary": "Cria um novo serviço",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.PrefRioServiceRequest"
              }
            }
          },
          "description": "Dados do serviço",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.PrefRioService"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/admin/tombamentos": {
      "get": {
        "description": "Lista tombamentos com paginação e filtros opcionais",
        "tags": [
          "tombamentos"
        ],
        "summary": "Lista tombamentos com paginação e filtros",
        "parameters": [
          {
            "description": "Página",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "description": "Resultados por página",
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "description": "Filtrar por origem (1746_v2_llm ou carioca-digital_v2_llm)",
            "name": "origem",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filtrar por criador",
            "name": "criado_por",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.TombamentoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Cria um mapeamento de serviço antigo para serviço novo na collection tombamentos_overlay",
        "tags": [
          "tombamentos"
        ],
        "summary": "Cria um novo tombamento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.TombamentoRequest"
              }
            }
          },
          "description": "Dados do tombamento",
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Tombamento"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/busca-hibrida-multi": {
      "get": {
        "description": "Realiza uma busca hibrida em varias colecoes combinando texto e embeddings",
        "tags": [
          "busca"
        ],
        "summary": "Busca hibrida em varias colecoes",
        "parameters": [
          {
            "description": "Lista de colecoes separadas por virgula",
            "name": "collections",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Termo de busca",
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Pagina",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "description": "Resultados por pagina",
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/categoria/{collections}": {
      "get": {
        "description": "Busca documentos de uma categoria especifica em uma ou multiplas colecoes retornando informacoes completas",
        "tags": [
          "busca"
        ],
        "summary": "Busca documentos por categoria",
        "parameters": [
          {
            "description": "Nome da colecao ou lista de colecoes separadas por virgula",
            "name": "collections",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Categoria dos documentos",
            "name": "categoria",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Pagina",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "description": "Resultados por pagina",
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/categorias-relevancia": {
      "get": {
        "description": "Retorna todas as categorias ordenadas por relevancia baseada na volumetria dos servicos",
        "tags": [
          "busca"
        ],
        "summary": "Busca categorias ordenadas por relevancia",
        "parameters": [
          {
            "description": "Lista de colecoes separadas por virgula",
            "name": "collections",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/documento/{collection}/{id}": {
      "get": {
        "description": "Busca um documento especifico por ID retornando todos os campos exceto embedding e campos normalizados",
        "tags": [
          "busca"
        ],
        "summary": "Busca documento por ID",
        "parameters": [
          {
            "description": "Nome da colecao",
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ID do documento",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://services.pref.rio/app-busca-search",
      "description": "Production server"
    },
    {
      "url": "https://services.staging.app.dados.rio/app-busca-search",
      "description": "Staging server"
    }
  ],
  "components": {
    "schemas": {
      "models.AgentsConfig": {
        "type": "object",
        "properties": {
          "exclusive_for_agents": {
            "type": "boolean"
          },
          "tool_hint": {
            "type": "string"
          }
        }
      },
      "models.Button": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string"
          },
          "is_enabled": {
            "type": "boolean"
          },
          "ordem": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "url_service": {
            "type": "string"
          }
        }
      },
...
      "models.PrefRioService": {
        "type": "object",
        "required": [
          "autor",
          "custo_servico",
          "descricao_completa",
          "nome_servico",
          "orgao_gestor",
          "resultado_solicitacao",
          "resumo",
          "tema_geral",
          "tempo_atendimento"
        ],
        "properties": {
          "agents": {
            "$ref": "#/components/schemas/models.AgentsConfig"
          },
          "autor": {
            "type": "string",
            "maxLength": 20000
          },
          "awaiting_approval": {
            "type": "boolean"
          },
          "buttons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.Button"
            }
          },
          "canais_digitais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canais_presenciais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "integer"
          },
          "custo_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "descricao_completa": {
            "type": "string",
            "maxLength": 20000
          },
          "documentos_necessarios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "embedding": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "extra_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "fixar_destaque": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "instrucoes_solicitante": {
            "type": "string",
            "maxLength": 20000
          },
          "is_free": {
            "type": "boolean"
          },
          "last_update": {
            "type": "integer"
          },
          "legislacao_relacionada": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nome_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "orgao_gestor": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          },
          "publico_especifico": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published_at": {
            "type": "integer"
          },
          "resultado_solicitacao": {
            "type": "string",
            "maxLength": 20000
          },
          "resumo": {
            "type": "string",
            "maxLength": 20000
          },
          "search_content": {
            "type": "string"
          },
          "servico_nao_cobre": {
            "type": "string",
            "maxLength": 20000
          },
          "status": {
            "description": "0=Draft, 1=Published",
            "type": "integer",
            "maximum": 1,
            "minimum": 0
          },
          "tema_geral": {
            "type": "string",
            "maxLength": 20000
          },
          "tempo_atendimento": {
            "type": "string",
            "maxLength": 20000
          }
        }
      },
      "models.PrefRioServiceRequest": {
        "type": "object",
        "required": [
          "nome_servico",
          "orgao_gestor",
          "publico_especifico",
          "resumo",
          "tema_geral"
        ],
        "properties": {
          "agents": {
            "$ref": "#/components/schemas/models.AgentsConfig"
          },
          "awaiting_approval": {
            "type": "boolean"
          },
          "buttons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.Button"
            }
          },
          "canais_digitais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canais_presenciais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "custo_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "descricao_completa": {
            "type": "string",
            "maxLength": 20000
          },
          "documentos_necessarios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "extra_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "fixar_destaque": {
            "type": "boolean"
          },
          "instrucoes_solicitante": {
            "type": "string",
            "maxLength": 20000
          },
          "is_free": {
            "type": "boolean"
          },
          "legislacao_relacionada": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nome_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "orgao_gestor": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          },
          "publico_especifico": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          },
          "published_at": {
            "type": "integer"
          },
          "resultado_solicitacao": {
            "type": "string",
            "maxLength": 20000
          },
          "resumo": {
            "type": "string",
            "maxLength": 20000
          },
          "servico_nao_cobre": {
            "type": "string",
            "maxLength": 20000
          },
          "status": {
            "type": "integer",
            "maximum": 1,
            "minimum": 0
          },
          "tema_geral": {
            "type": "string",
            "maxLength": 20000
          },
          "tempo_atendimento": {
            "type": "string",
            "maxLength": 20000
          }
        }
      },
      "models.PrefRioServiceResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "integer"
          },
          "out_of": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.PrefRioService"
            }
          }
        }
      },
...
      "models.Tombamento": {
        "type": "object",
        "required": [
          "criado_por",
          "id_servico_antigo",
          "id_servico_novo",
          "origem"
        ],
        "properties": {
          "criado_em": {
            "type": "integer"
          },
          "criado_por": {
            "type": "string",
            "maxLength": 20000
          },
          "id": {
            "type": "string"
          },
          "id_servico_antigo": {
            "type": "string",
            "maxLength": 20000
          },
          "id_servico_novo": {
            "type": "string",
            "maxLength": 20000
          },
          "observacoes": {
            "type": "string",
            "maxLength": 20000
          },
          "origem": {
            "type": "string",
            "enum": [
              "1746_v2_llm",
              "carioca-digital_v2_llm"
            ]
          }
        }
      },
      "models.TombamentoRequest": {
        "type": "object",
        "required": [
          "id_servico_antigo",
          "id_servico_novo",
          "origem"
        ],
        "properties": {
          "id_servico_antigo": {
            "type": "string",
            "maxLength": 20000
          },
          "id_servico_novo": {
            "type": "string",
            "maxLength": 20000
          },
          "observacoes": {
            "type": "string",
            "maxLength": 20000
          },
          "origem": {
            "type": "string",
            "enum": [
              "1746_v2_llm",
              "carioca-digital_v2_llm"
            ]
          }
        }
      },
      "models.TombamentoResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "integer"
          },
          "out_of": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "tombamentos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.Tombamento"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}